refactor(notification): refactor notification code
