+++ tests/ozpcenter_api/test_api_notification.py
@@ -10,6 +10,7 @@ from django.test import override_settings
 from rest_framework import status
 from rest_framework.test import APITestCase
 
+from tests.ozpcenter.helper import unittest_request_helper
 from ozpcenter import model_access as generic_model_access
 from ozpcenter.scripts import sample_data_generator as data_gen
 
@@ -25,7 +26,9 @@ class NotificationApiTest(APITestCase):
         """
         setUp is invoked before each test method
         """
-        self
+        self.expected_error = {'detail': 'Permission denied.',
+                               'error': True,
+                               'message': 'Only Stewards can delete notifications'}
 
     @classmethod
     def setUpTestData(cls):
@@ -34,27 +37,20 @@ class NotificationApiTest(APITestCase):
         """
         data_gen.run()
 
+    def _format_notification_response(self, response):
+        return ['{}-{}-{}'.format(entry['entity_id'], entry['notification_type'], ''.join(entry['message'].split())) for entry in response.data]
+
     def test_get_self_notification(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/self/notification/'
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
         for current_notification in response.data:
-            self.assertIn('id', current_notification)
-            self.assertIn('created_date', current_notification)
-            self.assertIn('expires_date', current_notification)
-            self.assertIn('message', current_notification)
-            self.assertIn('author', current_notification)
-            self.assertIn('listing', current_notification)
-            self.assertIn('agency', current_notification)
-            self.assertIn('notification_type', current_notification)
-            self.assertIn('peer', current_notification)
-            self.assertIn('notification_id', current_notification)
-            self.assertIn('read_status', current_notification)
-            self.assertIn('acknowledged_status', current_notification)
+            valid_keys = ['id', 'created_date', 'expires_date', 'message', 'author',
+                          'listing', 'agency', 'notification_type', 'peer',
+                          'notification_id', 'read_status', 'acknowledged_status']
+
+            for key in valid_keys:
+                self.assertIn(key, current_notification)
 
     def test_get_self_notification_unauthorized(self):
         url = '/api/self/notification/'
@@ -63,15 +59,10 @@ class NotificationApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
 
     def test_get_self_notification_ordering(self):
-        # Get default
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/self/notification/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
-        notification_list = ['{}-{}-{}'.format(entry['entity_id'], entry['notification_type'], ''.join(entry['message'].split())) for entry in response.data]
+        notification_list = self._format_notification_response(response)
 
         expected_data = ['147-listing-Saturnupdatenextweek',
                          '101-listing-Mallratsupdatenextweek',
@@ -124,34 +115,23 @@ class NotificationApiTest(APITestCase):
         self.assertListEqual(notification_list, expected_data)
 
         # Get reversed order
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/self/notification/?ordering=-notification__created_date'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
-        notification_list = ['{}-{}-{}'.format(entry['entity_id'], entry['notification_type'], ''.join(entry['message'].split())) for entry in response.data]
-
+        notification_list = self._format_notification_response(response)
         self.assertEqual(notification_list, expected_data)
 
         # Get ascending order
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/self/notification/?ordering=notification__created_date'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
-        notification_list = ['{}-{}-{}'.format(entry['entity_id'], entry['notification_type'], ''.join(entry['message'].split())) for entry in response.data]
+        notification_list = self._format_notification_response(response)
         self.assertEqual(notification_list, list(reversed(expected_data)))
 
     def test_dismiss_self_notification(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/self/notification/'
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
+
         mailbox_ids = []
         notification_ids = []
         for i in response.data:
@@ -209,76 +189,24 @@ class NotificationApiTest(APITestCase):
         self.assertEqual(expected, notification_ids)
 
         # now dismiss the first notification
-        dismissed_mailbox_id = mailbox_ids[0]
-        url = url + str(dismissed_mailbox_id) + '/'
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        url = '{}{}/'.format(url, mailbox_ids[0])
+        response = unittest_request_helper(self, url, 'DELETE', username='wsmith', status_code=204)
 
         # now get our notifications again, make sure the one was removed
         url = '/api/self/notification/'
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
+
         notification_ids = []
         for i in response.data:
             notification_ids.append([i['notification_id'], ''.join(i['message'].split())])
 
-        expected = [[222, 'Mallratsupdatenextweek'],
-                    [221, 'Mallratsupdatenextweek'],
-                    [213, 'JeanGreyupdatenextweek'],
-                    [212, 'IronManupdatenextweek'],
-                    [205, 'Grandfatherclockupdatenextweek'],
-                    [200, 'Diamondupdatenextweek'],
-                    [196, 'BreadBasketupdatenextweek'],
-                    [195, 'BreadBasketupdatenextweek'],
-                    [191, 'BaltimoreRavensupdatenextweek'],
-                    [190, 'BaltimoreRavensupdatenextweek'],
-                    [189, 'Azerothupdatenextweek'],
-                    [188, 'Azerothupdatenextweek'],
-                    [187, 'AirMailupdatenextweek'],
-                    [186, 'AirMailupdatenextweek'],
-                    [161, 'Auserhasratedlisting<b>Strokeplay</b>3stars'],
-                    [149, 'Auserhasratedlisting<b>Ruby</b>5stars'],
-                    [144, 'Auserhasratedlisting<b>ProjectManagement</b>1star'],
-                    [143, 'Auserhasratedlisting<b>ProjectManagement</b>2stars'],
-                    [125, 'Auserhasratedlisting<b>Moonshine</b>2stars'],
-                    [124, 'Auserhasratedlisting<b>Moonshine</b>5stars'],
-                    [101, 'Auserhasratedlisting<b>LocationLister</b>4stars'],
-                    [99, 'Auserhasratedlisting<b>Lager</b>2stars'],
-                    [98, 'Auserhasratedlisting<b>Lager</b>5stars'],
-                    [94, 'Auserhasratedlisting<b>KomodoDragon</b>1star'],
-                    [89, 'Auserhasratedlisting<b>JotSpot</b>4stars'],
-                    [77, 'Auserhasratedlisting<b>HouseTargaryen</b>5stars'],
-                    [76, 'Auserhasratedlisting<b>HouseStark</b>4stars'],
-                    [75, 'Auserhasratedlisting<b>HouseStark</b>1star'],
-                    [70, 'Auserhasratedlisting<b>Harley-DavidsonCVO</b>3stars'],
-                    [47, 'Auserhasratedlisting<b>ChartCourse</b>5stars'],
-                    [46, 'Auserhasratedlisting<b>ChartCourse</b>2stars'],
-                    [45, 'Auserhasratedlisting<b>BusinessManagementSystem</b>2stars'],
-                    [44, 'Auserhasratedlisting<b>BusinessManagementSystem</b>4stars'],
-                    [43, 'Auserhasratedlisting<b>BusinessManagementSystem</b>3stars'],
-                    [40, 'Auserhasratedlisting<b>BreadBasket</b>5stars'],
-                    [39, 'Auserhasratedlisting<b>BreadBasket</b>2stars'],
-                    [35, 'Auserhasratedlisting<b>Bleach</b>5stars'],
-                    [34, 'Auserhasratedlisting<b>Bleach</b>4stars'],
-                    [29, 'Auserhasratedlisting<b>BassFishing</b>4stars'],
-                    [23, 'Auserhasratedlisting<b>Barbecue</b>5stars'],
-                    [11, 'Auserhasratedlisting<b>AirMail</b>4stars'],
-                    [10, 'Auserhasratedlisting<b>AirMail</b>1star'],
-                    [9, 'Auserhasratedlisting<b>AirMail</b>3stars'],
-                    [8, 'Auserhasratedlisting<b>AirMail</b>5stars'],
-                    [2, 'Systemwillbefunctioninginadegredadedstatebetween1800Z-0400ZonA/B'],
-                    [1, 'Systemwillbegoingdownforapproximately30minutesonX/Yat1100Z']]
-
-        self.assertEqual(expected, notification_ids)
+        self.assertEqual(expected[1:], notification_ids)
 
     def test_get_pending_notifications(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/pending/'
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
         expires_at = [i['expires_date'] for i in response.data]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(len(expires_at) > 1)
         now = datetime.datetime.now(pytz.utc)
         for i in expires_at:
@@ -286,54 +214,38 @@ class NotificationApiTest(APITestCase):
             self.assertTrue(test_time > now)
 
     def test_get_pending_notifications_user_unauthorized(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/pending/'
-        response = self.client.get(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        unittest_request_helper(self, url, 'GET', username='jones', status_code=403)
 
     def test_all_pending_notifications_listing_filter(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/pending/?listing=1'  # ID 1 belongs to AcousticGuitar
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='bigbrother', status_code=200)
 
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        notification_list = self._format_notification_response(response)
+        notification_expires_date_list = [entry['expires_date'] for entry in response.data]
 
-        notification_list = [['{}-{}-{}'.format(entry['entity_id'], entry['notification_type'], ''.join(entry['message'].split())), entry['expires_date']] for entry in response.data]
         expected = ['1-listing-AcousticGuitarupdatenextweek',
                     '1-listing-Auserhasratedlisting<b>AcousticGuitar</b>5stars',
                     '1-listing-Auserhasratedlisting<b>AcousticGuitar</b>1star',
                     '1-listing-Auserhasratedlisting<b>AcousticGuitar</b>3stars']
 
-        self.assertEqual(expected, [entry[0] for entry in notification_list])
+        self.assertEqual(expected, notification_list)
 
         now = datetime.datetime.now(pytz.utc)
-        expires_at = [entry[1] for entry in notification_list]
-        for i in expires_at:
+        for i in notification_expires_date_list:
             test_time = datetime.datetime.strptime(i, "%Y-%m-%dT%H:%M:%S.%fZ").replace(tzinfo=pytz.utc)
             self.assertTrue(test_time > now)
 
     def test_all_pending_notifications_listing_filter_user_unauthorized(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/pending/?listing=1'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        unittest_request_helper(self, url, 'GET', username='jones', status_code=403)
+        # TODO Check response
 
     def test_get_expired_notifications(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/expired/'
-        response = self.client.get(url, format='json')
+        response = unittest_request_helper(self, url, 'GET', username='wsmith', status_code=200)
 
         expires_at = [i['expires_date'] for i in response.data]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(len(expires_at) > 1)
         now = datetime.datetime.now(pytz.utc)
         for i in expires_at:
@@ -342,26 +254,22 @@ class NotificationApiTest(APITestCase):
             self.assertTrue(test_time < now)
 
     def test_get_expired_notifications_user_unauthorized(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/expired/'
-        response = self.client.get(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        unittest_request_helper(self, url, 'GET', username='jones', status_code=403)
+        # TODO Check response
 
     # TODO should work when data script gets refactored (rivera 20160620)
     @skip("should work when data script gets refactored (rivera 20160620)")
     def test_all_expired_notifications_listing_filter(self):
+        url = '/api/notifications/expired/?listing=1'
+
         user = generic_model_access.get_profile('bigbrother').user
         self.client.force_authenticate(user=user)
-
-        url = '/api/notifications/expired/?listing=1'
         response = self.client.get(url, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
 
         ids = [i['id'] for i in response.data]
         expires_at = [i['expires_date'] for i in response.data]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(ids, [1])
         self.assertTrue(len(expires_at) == 1)
         now = datetime.datetime.now(pytz.utc)
@@ -371,83 +279,70 @@ class NotificationApiTest(APITestCase):
             self.assertTrue(test_time < now)
 
     def test_all_expired_notifications_listing_filter_user_unauthorized(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notifications/expired/?listing=1'
-        response = self.client.get(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        unittest_request_helper(self, url, 'GET', username='jones', status_code=403)
 
     # TODO: test_all_notifications_listing_filter (rivera 20160617)
 
     def test_create_system_notification(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         data = {'expires_date': '2016-09-01T15:45:55.322421Z',
                 'message': 'a simple test'}
-
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
         self.assertEqual(response.data['message'], 'a simple test')
         self.assertEqual(response.data['notification_type'], 'system')
 
     def test_create_system_notification_unauthorized_user(self):
         # test_create_system_notification_unauthorized_user
         # test unauthorized user - only org stewards and above can create
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         data = {'expires_date': '2016-09-01T15:45:55.322421Z',
                 'message': 'a simple test'}
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('jones').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_update_system_notification(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now)}
         url = '/api/notification/1/'
-        response = self.client.put(url, data, format='json')
 
+        user = generic_model_access.get_profile('wsmith').user
+        self.client.force_authenticate(user=user)
+        response = self.client.put(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         # TODO: Verify expires_date
 
     def test_update_system_notification_unauthorized_user(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now)}
         url = '/api/notification/1/'
-        response = self.client.put(url, data, format='json')
 
+        user = generic_model_access.get_profile('jones').user
+        self.client.force_authenticate(user=user)
+        response = self.client.put(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     # TODO below test should work when permission gets refactored (rivera 20160620)
     @skip("should work permissions gets refactored (rivera 20160620)")
     def test_update_system_notification_unauthorized_org_steward(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now)}
         url = '/api/notification/1/'
-        response = self.client.put(url, data, format='json')
 
+        user = generic_model_access.get_profile('wsmith').user
+        self.client.force_authenticate(user=user)
+        response = self.client.put(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_create_listing_notification_app_mall_steward(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -455,9 +350,12 @@ class NotificationApiTest(APITestCase):
             'id': 1
             }}
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
         self.assertEqual(response.data['message'], 'a simple listing test')
         self.assertEqual(response.data['notification_type'], 'listing')
         self.assertEqual(response.data['listing']['id'], 1)
@@ -465,9 +363,6 @@ class NotificationApiTest(APITestCase):
         self.assertTrue('expires_date' in data)
 
     def test_create_listing_notification_app_mall_steward_invalid_format(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -475,15 +370,15 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ['Valid Listing ID is required'])
 
     def test_create_listing_notification_app_mall_steward_invalid_id(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -491,15 +386,15 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ['Could not find listing'])
 
     def test_create_listing_agency_notification_app_mall_steward_invalid(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -508,15 +403,15 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ["Notifications can only be one type. Input: ['listing', 'agency']"])
 
     def test_create_listing_notification_org_steward(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -524,9 +419,12 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('wsmith').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
         self.assertEqual(response.data['message'], 'a simple listing test')
         self.assertEqual(response.data['notification_type'], 'listing')
         self.assertEqual(response.data['listing']['id'], 1)
@@ -534,9 +432,6 @@ class NotificationApiTest(APITestCase):
         self.assertTrue('expires_date' in data)
 
     def test_create_listing_notification_org_steward_invalid_format(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -544,15 +439,15 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('wsmith').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ["Valid Listing ID is required"])
 
     def test_create_listing_notification_org_steward_invalid_id(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple listing test',
@@ -560,27 +455,29 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('wsmith').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ["Could not find listing"])
 
     # TODO: test_create_listing_notification_org_steward_invalid (rivera 20160617)
     # TODO: test_create_listing_notification_user_unauthorized (rivera 20160617)
 
     def test_create_agency_notification_app_mall_steward(self):
-        url = '/api/notification/'
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'A Simple Agency Test',
-                'agency': {'id': 1}
-                }
+                'agency': {'id': 1}}
+        url = '/api/notification/'
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
         response = self.client.post(url, data, format='json')
-
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
         self.assertEqual(response.data['message'], 'A Simple Agency Test')
         self.assertEqual(response.data['notification_type'], 'agency')
         self.assertEqual(response.data['agency']['id'], 1)
@@ -588,33 +485,33 @@ class NotificationApiTest(APITestCase):
         self.assertTrue('expires_date' in data)
 
     def test_create_agency_notification_app_mall_steward_invalid_format(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple agency test',
                 'agency': {'invalid': 1}
                 }
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ['Valid Agency ID is required'])
 
     def test_create_agency_notification_app_mall_steward_invalid_id(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {'expires_date': str(now),
                 'message': 'a simple agency test',
                 'agency': {'id': -1}
                 }
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
         self.assertEqual(response.data['non_field_errors'], ['Could not find agency'])
 
     # TODO: test_create_agency_notification_org_steward (rivera 20160617)
@@ -622,9 +519,6 @@ class NotificationApiTest(APITestCase):
     # TODO: test_create_agency_notification_user_unauthorized (rivera 20160617)
 
     def test_create_peer_notification_app_mall_steward(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         now = datetime.datetime.now(pytz.utc)
         data = {"expires_date": str(now),
                 "message": "A Simple Peer to Peer Notification",
@@ -635,9 +529,12 @@ class NotificationApiTest(APITestCase):
                 }
 
         url = '/api/notification/'
-        response = self.client.post(url, data, format='json')
 
+        user = generic_model_access.get_profile('bigbrother').user
+        self.client.force_authenticate(user=user)
+        response = self.client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
         self.assertEqual(response.data['message'], 'A Simple Peer to Peer Notification')
         self.assertEqual(response.data['notification_type'], 'peer')
         self.assertEqual(response.data['agency'], None)
@@ -672,6 +569,36 @@ class NotificationApiTest(APITestCase):
     # TODO test_create_peer_notification_invalid (rivera 20160617)
     # TODO test_create_peer_bookmark_notification (rivera 20160617)
 
+    def _compare_library(self, usernames_list):
+        usernames_list_actual = {}
+        for username, ids_list in usernames_list.items():
+            user = generic_model_access.get_profile(username).user
+            self.client.force_authenticate(user=user)
+
+            url = '/api/self/library/'
+            response = self.client.get(url, format='json')
+            before_notification_ids = ['{}-{}'.format(entry['listing']['title'], entry['folder']) for entry in response.data]
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            usernames_list_actual[username] = before_notification_ids
+
+        for username, ids_list in usernames_list.items():
+            before_notification_ids = usernames_list_actual[username]
+            self.assertEqual(sorted(ids_list), sorted(before_notification_ids), 'Checking for {}'.format(username))
+
+    def _compare_user_notification(self, notification_user_list):
+        usernames_list = notification_user_list
+
+        for username, ids_list in usernames_list.items():
+            user = generic_model_access.get_profile(username).user
+            self.client.force_authenticate(user=user)
+
+            url = '/api/self/notification/'
+            response = self.client.get(url, format='json')
+
+            before_notification_ids = ['{}-{}'.format(entry.get('notification_type'), ''.join(entry.get('message').split())) for entry in response.data]
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            self.assertEqual(ids_list, before_notification_ids, 'Comparing Notifications for {}'.format(username))
+
     def test_create_peer_bookmark_notification_integration(self):
         """
         test_create_peer_bookmark_notification_integration
@@ -683,11 +610,52 @@ class NotificationApiTest(APITestCase):
         julia (minitrue, stewarded_orgs: minitrue, miniluv)
         bigbrother2 - minitrue
         """
+        # Library for users
+        user_library = {'bigbrother': ['Tornado-Weather',
+                                       'Lightning-Weather',
+                                       'Snow-Weather',
+                                       'Wolf Finder-Animals',
+                                       'Killer Whale-Animals',
+                                       'Lion Finder-Animals',
+                                       'Monkey Finder-Animals',
+                                       'Parrotlet-Animals',
+                                       'White Horse-Animals',
+                                       'Electric Guitar-Instruments',
+                                       'Acoustic Guitar-Instruments',
+                                       'Sound Mixer-Instruments',
+                                       'Electric Piano-Instruments',
+                                       'Piano-Instruments',
+                                       'Violin-Instruments',
+                                       'Bread Basket-None',
+                                       'Informational Book-None',
+                                       'Stop sign-None',
+                                       'Chain boat navigation-None',
+                                       'Gallery of Maps-None',
+                                       'Chart Course-None'],
+                        'jones': ['Bass Fishing-None', 'Killer Whale-None', 'Lager-None'],
+                        'julia': [],
+                        'wsmith': ['Air Mail-old',
+                                   'Bread Basket-old',
+                                   'Diamond-None',
+                                   'Grandfather clock-None',
+                                   'Baltimore Ravens-None',
+                                   'Iron Man-heros',
+                                   'Jean Grey-heros',
+                                   'Mallrats-heros',
+                                   'Azeroth-planets',
+                                   'Saturn-planets']}
+
+        # Create Bookmark
         response = _create_create_bookmark(self, 'wsmith', 3, folder_name='foldername1', status_code=201)
         self.assertEqual(response.data['listing']['id'], 3)
+        user_library['wsmith'].append('{}-{}'.format(response.data['listing']['title'], response.data['folder']))
+        self._compare_library(user_library)
 
+        # Create Bookmark
         response = _create_create_bookmark(self, 'wsmith', 4, folder_name='foldername1', status_code=201)
         self.assertEqual(response.data['listing']['id'], 4)
+        user_library['wsmith'].append('{}-{}'.format(response.data['listing']['title'], response.data['folder']))
+        self._compare_library(user_library)
 
         # Compare Notifications for users
         usernames_list = {'bigbrother': ['listing-WolfFinderupdatenextweek',
@@ -906,110 +874,13 @@ class NotificationApiTest(APITestCase):
             before_notification_ids = usernames_list_actual[username]
             self.assertEqual(ids_list, before_notification_ids, 'Checking for {}'.format(username))
 
-        # Compare Library for users
-        usernames_list = {'bigbrother': ['Tornado-Weather',
-                                        'Lightning-Weather',
-                                        'Snow-Weather',
-                                        'Wolf Finder-Animals',
-                                        'Killer Whale-Animals',
-                                        'Lion Finder-Animals',
-                                        'Monkey Finder-Animals',
-                                        'Parrotlet-Animals',
-                                        'White Horse-Animals',
-                                        'Electric Guitar-Instruments',
-                                        'Acoustic Guitar-Instruments',
-                                        'Sound Mixer-Instruments',
-                                        'Electric Piano-Instruments',
-                                        'Piano-Instruments',
-                                        'Violin-Instruments',
-                                        'Bread Basket-None',
-                                        'Informational Book-None',
-                                        'Stop sign-None',
-                                        'Chain boat navigation-None',
-                                        'Gallery of Maps-None',
-                                        'Chart Course-None'],
-                          'jones': ['Bass Fishing-None', 'Killer Whale-None', 'Lager-None'],
-                          'julia': [],
-                          'wsmith': ['Air Mail-old',
-                                     'Albatron Technology-foldername1',
-                                     'Aliens-foldername1',
-                                     'Bread Basket-old',
-                                     'Diamond-None',
-                                     'Grandfather clock-None',
-                                     'Baltimore Ravens-None',
-                                     'Iron Man-heros',
-                                     'Jean Grey-heros',
-                                     'Mallrats-heros',
-                                     'Azeroth-planets',
-                                     'Saturn-planets']}
-
-        usernames_list_actual = {}
-        for username, ids_list in usernames_list.items():
-            user = generic_model_access.get_profile(username).user
-            self.client.force_authenticate(user=user)
-
-            url = '/api/self/library/'
-            response = self.client.get(url, format='json')
-            before_notification_ids = ['{}-{}'.format(entry['listing']['title'], entry['folder']) for entry in response.data]
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-            usernames_list_actual[username] = before_notification_ids
-
-        for username, ids_list in usernames_list.items():
-            before_notification_ids = usernames_list_actual[username]
-            self.assertEqual(ids_list, before_notification_ids, 'Checking for {}'.format(username))
+        self._compare_library(user_library)
 
         # Create Bookmark
         response = _create_create_bookmark(self, 'bigbrother', 4, folder_name='foldername2', status_code=201)
         self.assertEqual(response.data['listing']['id'], 4)
-
-        # Compare Library for users
-        usernames_list = {'wsmith': ['Air Mail-old',
-                                     'Albatron Technology-foldername1',
-                                     'Aliens-foldername1',
-                                     'Bread Basket-old',
-                                     'Diamond-None',
-                                     'Grandfather clock-None',
-                                     'Baltimore Ravens-None',
-                                     'Iron Man-heros',
-                                     'Jean Grey-heros',
-                                     'Mallrats-heros',
-                                     'Azeroth-planets',
-                                     'Saturn-planets'],
-                          'julia': [],
-                          'jones': ['Bass Fishing-None', 'Killer Whale-None', 'Lager-None'],
-                          'bigbrother': ['Tornado-Weather',
-                                         'Aliens-foldername2',
-                                         'Lightning-Weather',
-                                         'Snow-Weather',
-                                         'Wolf Finder-Animals',
-                                         'Killer Whale-Animals',
-                                         'Lion Finder-Animals',
-                                         'Monkey Finder-Animals',
-                                         'Parrotlet-Animals',
-                                         'White Horse-Animals',
-                                         'Electric Guitar-Instruments',
-                                         'Acoustic Guitar-Instruments',
-                                         'Sound Mixer-Instruments',
-                                         'Electric Piano-Instruments',
-                                         'Piano-Instruments',
-                                         'Violin-Instruments',
-                                         'Bread Basket-None',
-                                         'Informational Book-None',
-                                         'Stop sign-None',
-                                         'Chain boat navigation-None',
-                                         'Gallery of Maps-None',
-                                         'Chart Course-None']}
-
-        for username, ids_list in usernames_list.items():
-            user = generic_model_access.get_profile(username).user
-            self.client.force_authenticate(user=user)
-
-            url = '/api/self/library/'
-            response = self.client.get(url, format='json')
-
-            before_notification_ids = ['{}-{}'.format(entry['listing']['title'], entry['folder']) for entry in response.data]
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-            self.assertEqual(before_notification_ids, ids_list, 'Comparing Library for {}'.format(username))
+        user_library['bigbrother'].append('{}-{}'.format(response.data['listing']['title'], response.data['folder']))
+        self._compare_library(user_library)
 
         # Create Bookmark Notification
         bookmark_notification_ids = []
@@ -1108,58 +979,29 @@ class NotificationApiTest(APITestCase):
                                          'Gallery of Maps-None',
                                          'Chart Course-None']}
 
-        for username, ids_list in usernames_list.items():
-            user = generic_model_access.get_profile(username).user
-            self.client.force_authenticate(user=user)
-
-            url = '/api/self/library/'
-            response = self.client.get(url, format='json')
-
-            before_notification_ids = ['{}-{}'.format(entry['listing']['title'], entry['folder']) for entry in response.data]
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-            self.assertEqual(before_notification_ids, ids_list, 'Comparing Library for {}'.format(username))
+        self._compare_library(usernames_list)
 
         # Compare Notifications for users
         usernames_list = copy.deepcopy(usernames_list_main)
         usernames_list['julia'] = bookmark_notification_ids[::-1] + usernames_list_main['julia']
 
-        for username, ids_list in usernames_list.items():
-            user = generic_model_access.get_profile(username).user
-            self.client.force_authenticate(user=user)
-
-            url = '/api/self/notification/'
-            response = self.client.get(url, format='json')
-
-            before_notification_ids = ['{}-{}'.format(entry.get('notification_type'), ''.join(entry.get('message').split())) for entry in response.data]
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-            self.assertEqual(ids_list, before_notification_ids, 'Comparing Notifications for {}'.format(username))
+        self._compare_user_notification(usernames_list)
 
     def test_delete_system_notification_apps_mall_steward(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notification/1/'
-        response = self.client.delete(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        unittest_request_helper(self, url, 'DELETE', username='bigbrother', status_code=204)
 
     # TODO below test should work when permission gets refactored (rivera 20160620)
     @skip("should work when permission gets refactored (rivera 20160620)")
     def test_delete_system_notification_org_steward(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
         url = '/api/notification/1/'
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        response = unittest_request_helper(self, url, 'DELETE', username='wsmith', status_code=403)
+        self.assertEqual(response.data, self.expected_error)
 
     def test_delete_system_notification_user_unauthorized(self):
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-
         url = '/api/notification/1/'
-        response = self.client.delete(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        response = unittest_request_helper(self, url, 'DELETE', username='jones', status_code=403)
+        self.assertEqual(response.data, self.expected_error)
 
     # TODO: Unittest for below
     # AMLNG-378 - As a user, I want to receive notification about changes on Listings I've bookmarked
